# YJIT: Skip padding jumps to side exits on Arm