This PR revisits the idea I shared at https://github.com/ruby/ruby/pull/6757#issuecomment-1319371923.

Jumps to side exits are non-patchable, so we don't need to generate nops for these cases.

## Benchmark
A release build on Arm becomes 1.02x faster.

```
before: ruby 3.2.0dev (2022-11-21T18:52:44Z master cf05c202ce) +YJIT [arm64-darwin22]
after: ruby 3.2.0dev (2022-11-22T19:09:31Z yjit-side-exit-ptr e513fe1ad3) +YJIT [arm64-darwin22]

----------  -----------  ----------  ----------  ----------  ------------  -------------
bench       before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
railsbench  710.1        2.4         693.6       1.1         1.02          0.99
----------  -----------  ----------  ----------  ----------  ------------  -------------
```

<details>

```
$ ./run_benchmarks.rb railsbench -e "before::/opt/rubies/yjit-release-before-$arch/bin/ruby --yjit" -e "after::/opt/rubies/yjit-release-after-$arch/bin/ruby --yjit"
Running benchmark "railsbench" (1/1)
/opt/rubies/yjit-release-before-arm64/bin/ruby --yjit -I ./harness benchmarks/railsbench/benchmark.rb
ruby 3.2.0dev (2022-11-21T18:52:44Z master cf05c202ce) +YJIT [arm64-darwin22]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
Command: bin/rails db:migrate db:seed
Deleted all 100 posts
Creating 100 posts....................................................................................................
itr #1: 857ms
itr #2: 730ms
itr #3: 704ms
itr #4: 739ms
itr #5: 723ms
itr #6: 715ms
itr #7: 719ms
itr #8: 699ms
itr #9: 680ms
itr #10: 719ms
itr #11: 698ms
itr #12: 699ms
itr #13: 706ms
itr #14: 718ms
itr #15: 694ms
itr #16: 727ms
itr #17: 705ms
itr #18: 697ms
itr #19: 686ms
itr #20: 715ms
itr #21: 740ms
itr #22: 703ms
itr #23: 694ms
itr #24: 731ms
itr #25: 696ms
Average of last 10, non-warmup iters: 710ms
Running benchmark "railsbench" (1/1)
/opt/rubies/yjit-release-after-arm64/bin/ruby --yjit -I ./harness benchmarks/railsbench/benchmark.rb
ruby 3.2.0dev (2022-11-22T19:09:31Z yjit-side-exit-ptr e513fe1ad3) +YJIT [arm64-darwin22]
Command: bundle check 2> /dev/null || bundle install
The Gemfile's dependencies are satisfied
Command: bin/rails db:migrate db:seed
Deleted all 100 posts
Creating 100 posts....................................................................................................
itr #1: 864ms
itr #2: 706ms
itr #3: 688ms
itr #4: 696ms
itr #5: 690ms
itr #6: 687ms
itr #7: 710ms
itr #8: 706ms
itr #9: 687ms
itr #10: 686ms
itr #11: 685ms
itr #12: 694ms
itr #13: 686ms
itr #14: 685ms
itr #15: 692ms
itr #16: 705ms
itr #17: 685ms
itr #18: 691ms
itr #19: 688ms
itr #20: 687ms
itr #21: 689ms
itr #22: 705ms
itr #23: 704ms
itr #24: 691ms
itr #25: 686ms
Average of last 10, non-warmup iters: 693ms
Total time spent benchmarking: 41s

before: ruby 3.2.0dev (2022-11-21T18:52:44Z master cf05c202ce) +YJIT [arm64-darwin22]
after: ruby 3.2.0dev (2022-11-22T19:09:31Z yjit-side-exit-ptr e513fe1ad3) +YJIT [arm64-darwin22]
end_time: 2022-11-22 11:12:18 PST (-0800)

----------  -----------  ----------  ----------  ----------  ------------  -------------
bench       before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
railsbench  710.1        2.4         693.6       1.1         1.02          0.99
----------  -----------  ----------  ----------  ----------  ------------  -------------
Legend:
- before/after: ratio of before/after time. Higher is better for after. Above 1 represents a speedup.
- after 1st itr: ratio of before/after time for the first benchmarking iteration.
```

</details>

## Code Size
`inline_code_size` is decreased by 8% with a release build on Arm.

### Before
```
inline_code_size:         4795980
outlined_code_size:       4794088
freed_code_size:                0
code_region_size:         9601024
```

### After
```
inline_code_size:         4415324
outlined_code_size:       4411532
freed_code_size:                0
code_region_size:         8830976
```